requirements.txt
customtkinter
Pillow
PyMySQL
pytest
pylint